+++ src/styles/global.css
@@ -33,7 +33,7 @@ body {
 
 a {
   text-decoration: none;
-  color: #1D86FF;
+  color: #56B4E9;
 }
 
 h1,h2,h3,h4,h5,h6 {
